requirements.txt
fastapi
uvicorn
torch
clip-by-openai
pillow
numpy
faiss-cpu
pymongo
